🔥 [remove] expo 보일러플레이트 코드 제거
